+++ internal/control-plane/translator/clusters.go
@@ -6,48 +6,31 @@ import (
 	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 
-	"github.com/solo-io/gloo/internal/control-plane/filewatcher"
-	"github.com/solo-io/gloo/internal/control-plane/reporter"
+	"github.com/solo-io/gloo/internal/control-plane/snapshot"
 	"github.com/solo-io/gloo/internal/control-plane/translator/defaults"
 	"github.com/solo-io/gloo/pkg/api/types/v1"
 	"github.com/solo-io/gloo/pkg/endpointdiscovery"
 	"github.com/solo-io/gloo/pkg/plugins"
-	"github.com/solo-io/gloo/pkg/secretwatcher"
 )
 
-// Clusters
-
-func (t *Translator) computeClusters(cfg *v1.Config, dependencies *pluginDependencies, endpoints endpointdiscovery.EndpointGroups) ([]*envoyapi.Cluster, []reporter.ConfigObjectReport) {
+func (t *Translator) computeClusters(inputs *snapshot.Cache, configErrs configErrors) []*envoyapi.Cluster {
 	var (
-		reports  []reporter.ConfigObjectReport
 		clusters []*envoyapi.Cluster
 	)
-	for _, upstream := range cfg.Upstreams {
-		_, edsCluster := endpoints[upstream.Name]
-		cluster, err := t.computeCluster(cfg, dependencies, upstream, edsCluster)
-		// only append valid clusters
-		if err == nil {
-			clusters = append(clusters, cluster)
+	for _, upstream := range inputs.Cfg.Upstreams {
+		cluster, err := t.computeCluster(inputs, upstream)
+		if err != nil {
+			configErrs.addError(upstream, err)
+			continue
 		}
-		reports = append(reports, createReport(upstream, err))
+		// only append valid clusters
+		clusters = append(clusters, cluster)
 	}
-	return clusters, reports
+	return clusters
 }
 
-func (t *Translator) computeCluster(cfg *v1.Config, dependencies *pluginDependencies, upstream *v1.Upstream, edsCluster bool) (*envoyapi.Cluster, error) {
-	out := &envoyapi.Cluster{
-		Name:     upstream.Name,
-		Metadata: new(envoycore.Metadata),
-	}
-	if edsCluster {
-		out.Type = envoyapi.Cluster_EDS
-	}
-
-	timeout := upstream.ConnectionTimeout
-	if timeout == 0 {
-		timeout = defaults.ClusterConnectionTimeout
-	}
-	out.ConnectTimeout = timeout
+func (t *Translator) computeCluster(inputs *snapshot.Cache, upstream *v1.Upstream) (*envoyapi.Cluster, error) {
+	out := initializeCluster(upstream, inputs.Endpoints)
 
 	var upstreamErrors error
 	for _, plug := range t.plugins {
@@ -55,13 +38,11 @@ func (t *Translator) computeCluster(cfg *v1.Config, dependencies *pluginDependen
 		if !ok {
 			continue
 		}
+		secrets, files := dependenciesForPlugin(inputs, upstreamPlugin)
 		params := &plugins.UpstreamPluginParams{
 			EnvoyNameForUpstream: clusterName,
-		}
-		deps := dependenciesForPlugin(cfg, upstreamPlugin, dependencies)
-		if deps != nil {
-			params.Secrets = deps.Secrets
-			params.Files = deps.Files
+			Secrets: secrets,
+			Files: files,
 		}
 
 		if err := upstreamPlugin.ProcessUpstream(params, upstream, out); err != nil {
@@ -74,6 +55,24 @@ func (t *Translator) computeCluster(cfg *v1.Config, dependencies *pluginDependen
 	return out, upstreamErrors
 }
 
+func initializeCluster(upstream *v1.Upstream, endpoints endpointdiscovery.EndpointGroups) *envoyapi.Cluster {
+	out := &envoyapi.Cluster{
+		Name:     upstream.Name,
+		Metadata: new(envoycore.Metadata),
+	}
+	// set Type = EDS if we have endpoints for the upstream
+	if _, edsCluster := endpoints[upstream.Name]; edsCluster {
+		out.Type = envoyapi.Cluster_EDS
+	}
+	// set connection timeout
+	timeout := upstream.ConnectionTimeout
+	if timeout == 0 {
+		timeout = defaults.ClusterConnectionTimeout
+	}
+	out.ConnectTimeout = timeout
+	return out
+}
+
 // TODO: add more validation here
 func validateCluster(c *envoyapi.Cluster) error {
 	if c.Type == envoyapi.Cluster_STATIC || c.Type == envoyapi.Cluster_STRICT_DNS || c.Type == envoyapi.Cluster_LOGICAL_DNS {
@@ -82,40 +81,4 @@ func validateCluster(c *envoyapi.Cluster) error {
 		}
 	}
 	return nil
-}
-
-func dependenciesForPlugin(cfg *v1.Config, plug plugins.TranslatorPlugin, dependencies *pluginDependencies) *pluginDependencies {
-	dependencyRefs := plug.GetDependencies(cfg)
-	if dependencyRefs == nil {
-		return nil
-	}
-	pluginDeps := &pluginDependencies{
-		Secrets: make(secretwatcher.SecretMap),
-		Files:   make(filewatcher.Files),
-	}
-	for _, ref := range dependencyRefs.SecretRefs {
-		item, ok := dependencies.Secrets[ref]
-		if ok {
-			pluginDeps.Secrets[ref] = item
-		}
-	}
-	for _, ref := range dependencyRefs.FileRefs {
-		item, ok := dependencies.Files[ref]
-		if ok {
-			pluginDeps.Files[ref] = item
-		}
-	}
-	return pluginDeps
-}
-
-func deduplicateClusters(clusters []*envoyapi.Cluster) []*envoyapi.Cluster {
-	mapped := make(map[string]bool)
-	var deduped []*envoyapi.Cluster
-	for _, c := range clusters {
-		if _, added := mapped[c.Name]; added {
-			continue
-		}
-		deduped = append(deduped, c)
-	}
-	return deduped
 }
