+++ internal/control-plane/translator/internal_types.go
