+++ internal/control-plane/translator/names.go
@@ -1,5 +1,9 @@
 package translator
 
+import (
+	"github.com/solo-io/gloo/pkg/api/types/v1"
+)
+
 // for future-proofing possible safety issues with bad upstream names
 func clusterName(upstreamName string) string {
 	return upstreamName
@@ -10,3 +14,6 @@ func virtualHostName(virtualServiceName string) string {
 	return virtualServiceName
 }
 
+func routeConfigName(listener *v1.Listener) string {
+	return listener.Name+"-routes"
+}
