+++ internal/control-plane/translator/trim.go
